requirements.txt
flask=>3.0.3
boto3>=1.34.4
pytest>=7.2.1
pypdf>=4.2.0
PyYAML~=6.0.1
flask~=3.0.3
